requirements.txt
asgiref==3.5.2
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.0.0
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.2.0
platformdirs==3.5.1
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.2.1
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
uritemplate==4.1.1
urllib3==2.0.2
